move DeployApi into parent class
